refactor(psl): split psl.rs into psl/ module with dat.rs
Pure code move with no behavior change. Splits the single-file PSL module
into a directory layout to make room for additional reader implementations.
The trait and MockPublicSuffixList stay in mod.rs; DatFilePublicSuffixList
moves to dat.rs.
